Input controllés - Pages et Composants - Terminés

+++ front/src/components/ModalSuppression/ModalSuppression.jsx
@@ -1,4 +1,4 @@
-import {React, useState, useContext} from 'react'
+import { React, useState, useContext } from 'react'
 import axios from "axios"
 import { toast } from 'react-toastify'
 import { AuthContext } from '../../context/AuthContext'
@@ -11,21 +11,23 @@ import modalcgv from "../ModalCGV/modalcgv.module.css"
 
 const ModalSuppression = () => {
 
-        const {auth, deconnexion} = useContext(AuthContext);
+    const { auth, deconnexion } = useContext(AuthContext);
 
-        const deleteUser = async () => {
+    const deleteUser = async () => {
+        if (URL.USER_DELETE) {
             try {
                 const response = await axios.delete(`${URL.USER_DELETE}/${auth._id}`, { withCredentials: true })
-                console.log(response)
+                // console.log(response)
                 if (response.status === 200) {
                     toast.success("Compte supprimé avec succès.", { autoClose: 2000 })
                     deconnexion();
-                    
+
                 }
             } catch (error) {
                 console.error("Erreur lors de la suppression du profil", error)
             }
         }
+    }
 
     const [modal, setModal] = useState(false)
 
@@ -49,13 +51,13 @@ const ModalSuppression = () => {
                                 <p>Toutes vos données seront définitvement supprimées.</p>
                                 <p>Pensez à récupérer toutes vos factures pour tout besoin éventuel.</p>
                             </div>
-                <div className={modalcgv.contientBtnSuppression}>
-                    <button className={modalcgv.btnRevenir} onClick={activeModal}>Non, je me suis trompé de bouton</button>
-                    <button className={modalcgv.btnSuppression} onClick={() => deleteUser(auth._id)}>Oui, je veux supprimer mon compte</button>
-                </div>
+                            <div className={modalcgv.contientBtnSuppression}>
+                                <button className={modalcgv.btnRevenir} onClick={activeModal}>Non, je me suis trompé de bouton</button>
+                                <button className={modalcgv.btnSuppression} onClick={() => deleteUser(auth._id)}>Oui, je veux supprimer mon compte</button>
+                            </div>
                         </div>
 
-                    </div> 
+                    </div>
                 </div>
             )}
         </div>
